build: Update versioneer from 0.15 to 0.21
One benefit of this update is to support the upcoming Python 3.12
release. Old versioneer versions use configparser.SafeConfigParser.
This was deprecated in Python 3.2, and is being removed in 3.12.

To do the update, I first manually removed the code appended by the
old version to the end of __init__.py, because the update does not
do that (instead it appends improved but similar *code* after it).
Then I followed the current instructions for a "Vendored mode"
installation, by installing the versioneer package in a temporary
virtual environment and running "versioneer install --vendor". I
did this from the top-level repository directory where setup.py is.

As expected, this reported that the old versioneer.py was replaced,
and that the effect of running the new script was to create
_version.py and append to __init__.py (in the package subdir).