tests: Add setuptools as a dev dependency
This is needed in Python 3.12, when using tox, for the
pkg_resources import in conftest.py to succeed.

The pkg_resources module is provided by setuptools, and apparently
setuptools (or a sufficient version) is not reliably installed
in Python 3.12 when using tox. Since the test code explicitly uses
a module that setuptools provides, I think it makes sense to just
consider setuptools to be a dev/test dependency and list it as
such.