workflows: Avoid duplication in another way
This lists all Python versions in the same list, given as the value
of the python-version key for the matrix, as was done in the past.
However, it avoids producing two separate experimental and
non-experimental jobs for the same version that is intended to be
experimental, by listing a general and overrridden "experimental"
key in "include", and *not* listing "experimental" outside
"include".

This is closer to the spirit of the old way, while avoiding
the unintended duplication. It is well-suited to longer term use
(though "3.12.0-rc.1 - 3.12" should still be changed to just "3.12"
sometime soon after the stable 3.12.0 is released), because it will
always be immediately clear upon inspection what Python versions
are being tested.